+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 
 ### jannovar-core
 
+* Renaming `*.getChange()` to `*.getGenomeVariant()`
+* Renaming `VariantAnnotator.buildAnnotationList` to `.buildAnnotations`,
+  `VariantContextAnnotator.buildAnnotationList` to `.buildAnnotations`,
+  and `VariantContextAnnotator.buildErrorAnnotationList` to
+  `VariantContextAnnotator.buildErrorAnnotations`
+* VariantAnnotations does not implement `List<Annotation>` any more
+* Adding `VariantAnnotations.getAnnotations`
+* Renaming `AnnotationList` to `VariantAnnotations`
 * changing treatment of insertions at exon/intron junctions; they are
   considered as intronic insertions now that affect splicing
 * converting `GenomeVariant` of `AnnotationList` to always be on the forward